cleanup in hyperiond zeroconf switchable between static and shared linking
